put_mem modifide

+++ printf/include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: junhseo <junhseo@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/18 14:54:26 by junhseo           #+#    #+#             */
-/*   Updated: 2023/04/07 19:03:12 by junhseo          ###   ########.fr       */
+/*   Updated: 2023/04/07 19:04:20 by junhseo          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 
 int		ft_printf(const char *str, ...);
 int		is_type(const char *type_specifier, const char str);
-void	put_sx(unsigned long n, int *count);
-void	put_lx(unsigned long n, int *count);
+void	put_sx(unsigned int n, int *count);
+void	put_lx(unsigned int n, int *count);
 void	put_u(unsigned int n, int *count);
 void	put_per(int *count);
 void	put_mem(unsigned long ptr, int *count);
